+++ src/main/java/roomescape/reservation/application/repository/ReservationTimeRepository.java
@@ -3,7 +3,7 @@ package roomescape.reservation.application.repository;
 import java.time.LocalTime;
 import java.util.List;
 import java.util.Optional;
-import roomescape.reservation.domain.aggregate.ReservationTime;
+import roomescape.reservation.domain.ReservationTime;
 
 public interface ReservationTimeRepository {
     ReservationTime insert(LocalTime reservationTime);
