+++ src/main/java/roomescape/reservation/presentation/dto/ReservationResponse.java
@@ -1,7 +1,7 @@
 package roomescape.reservation.presentation.dto;
 
 import java.time.LocalDate;
-import roomescape.reservation.domain.aggregate.Reservation;
+import roomescape.reservation.domain.Reservation;
 
 public class ReservationResponse {
     private Long id;
