+++ src/main/java/roomescape/reservation/presentation/dto/ReservationTimeResponse.java
@@ -1,7 +1,7 @@
 package roomescape.reservation.presentation.dto;
 
 import java.time.LocalTime;
-import roomescape.reservation.domain.aggregate.ReservationTime;
+import roomescape.reservation.domain.ReservationTime;
 
 public class ReservationTimeResponse {
     private Long id;
