+++ src/test/java/roomescape/reservation/infrastructure/dao/ReservationDaoTest.java
@@ -13,8 +13,8 @@ import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import roomescape.global.exception.DeleteReservationException;
 import roomescape.reservation.application.dto.CreateReservationRequest;
-import roomescape.reservation.domain.aggregate.ReservationDate;
-import roomescape.reservation.domain.aggregate.ReservationName;
+import roomescape.reservation.domain.ReservationDate;
+import roomescape.reservation.domain.ReservationName;
 import roomescape.reservation.presentation.dto.ReservationTimeRequest;
 
 @JdbcTest
